+++ lib/InputTest.ts
@@ -1,5 +1,5 @@
 import { EventEmitter } from 'events';
-import { TestNames } from './constants';
+import { TestName, WarningName } from './constants';
 import {
   AlreadyStoppedError,
   DiagnosticError,
@@ -14,7 +14,7 @@ import {
   getUserMedia,
   GetUserMediaUnsupportedError,
 } from './polyfills';
-import { SubsetRequired, TimeMeasurement } from './types';
+import { SubsetRequired, TimeMeasurement, VolumeStats } from './types';
 import { detectSilence } from './utils';
 import {
   InvalidityRecord,
@@ -57,6 +57,28 @@ export declare interface InputTest {
     event: InputTest.Events.Volume,
     value: number,
   ): boolean;
+  /**
+   * This event is emitted when the test encounters a non-fatal warning during
+   * its run-time.
+   * @param event [[InputTest.Events.Warning]]
+   * @param warning The warning that the test encountered.
+   * @private
+   */
+  emit(
+    event: InputTest.Events.Warning,
+    warning: WarningName,
+  ): boolean;
+  /**
+   * This event is emitted when the test clears a previously encountered
+   * non-fatal warning during its run-time.
+   * @param event [[InputTest.Events.WarningCleared]]
+   * @param warning The warning that the test encountered that should be cleared.
+   * @private
+   */
+  emit(
+    event: InputTest.Events.WarningCleared,
+    warning: WarningName,
+  ): boolean;
 
   /**
    * Raised upon completion of the test.
@@ -97,6 +119,32 @@ export declare interface InputTest {
     event: InputTest.Events.Volume,
     listener: (value: number) => any,
   ): this;
+  /**
+   * Raised by the test when the test encounters a non-fatal warning during
+   * its run-time.
+   * @param event [[InputTest.Events.Warning]]
+   * @param listener A callback that expects the following parameters:
+   *  A [[DiagnosticWarning]].
+   * @returns This [[InputTest]] instance.
+   * @event
+   */
+  on(
+    event: InputTest.Events.Warning,
+    listener: (warning: WarningName) => any,
+  ): this;
+  /**
+   * Raised by the test when the test clears a previously encountered non-fatal
+   * warning during its run-time.
+   * @param event [[InputTest.Events.WarningCleared]]
+   * @param listener A callback that expects the following parameters:
+   *  A [[DiagnosticWarning]] name.
+   * @returns This [[InputTest]] instance.
+   * @event
+   */
+  on(
+    event: InputTest.Events.WarningCleared,
+    listener: (warning: WarningName) => any,
+  ): this;
 }
 
 /**
@@ -109,7 +157,7 @@ export class InputTest extends EventEmitter {
   /**
    * Name of the test.
    */
-  static testName: TestNames.InputAudioDevice = TestNames.InputAudioDevice;
+  static testName: TestName.InputAudioDevice = TestName.InputAudioDevice;
 
   /**
    * Default options for the `InputTest`.
@@ -122,6 +170,11 @@ export class InputTest extends EventEmitter {
     getUserMedia,
     volumeEventIntervalMs: 100,
   };
+
+  /**
+   * Active warnings to keep track of.
+   */
+  readonly activeWarnings: Set<WarningName> = new Set();
 
   /**
    * An `AudioContext` to use for generating volume levels.
@@ -148,10 +201,6 @@ export class InputTest extends EventEmitter {
    */
   private readonly _errors: DiagnosticError[] = [];
   /**
-   * The maximum volume level from the audio source.
-   */
-  private _maxValue: number = 0;
-  /**
    * A `MediaStream` that is created from the input device.
    */
   private _mediaStream: MediaStream | null = null;
@@ -168,7 +217,10 @@ export class InputTest extends EventEmitter {
    * Volume levels generated from the audio source during the run time of the
    * test.
    */
-  private readonly _values: number[] = [];
+  private readonly _volumeStats: VolumeStats = {
+    timestamps: [],
+    values: [],
+  };
   /**
    * The timeout that causes the volume event to loop; created by `setTimeout`.
    */
@@ -194,7 +246,7 @@ export class InputTest extends EventEmitter {
    * override the result from determining whether audio is silent from the collected volume levels.
    */
   stop(pass: boolean = true): InputTest.Report | undefined {
-    if (this._endTime) {
+    if (typeof this._endTime === 'number') {
       this._onWarning(new AlreadyStoppedError());
       return;
     }
@@ -203,7 +255,7 @@ export class InputTest extends EventEmitter {
     this._cleanup();
 
     this._endTime = Date.now();
-    const didPass: boolean = pass && !detectSilence(this._values);
+    const didPass: boolean = pass && !detectSilence(this._volumeStats.values);
     const report: InputTest.Report = {
       deviceId: this._options.deviceId || (
         this._defaultDevices.audioinput &&
@@ -212,7 +264,7 @@ export class InputTest extends EventEmitter {
       didPass,
       errors: this._errors,
       testName: InputTest.testName,
-      values: this._values,
+      values: this._volumeStats.values,
     };
 
     if (this._startTime) {
@@ -266,11 +318,57 @@ export class InputTest extends EventEmitter {
    * @param value the volume
    */
   private _onVolume(value: number): void {
-    if (value > this._maxValue) {
-      this._maxValue = value;
+    const now = Date.now();
+
+    if (!this._volumeStats.max || value > this._volumeStats.max) {
+      this._volumeStats.max = value;
     }
-    this._values.push(value);
+    this._volumeStats.values.push(value);
+    this._volumeStats.timestamps.push(now);
     this.emit(InputTest.Events.Volume, value);
+
+    // Find the last 3 seconds worth of volume values.
+    const startIndex = this._volumeStats.timestamps.findIndex(
+      (timestamp: number) => now - timestamp <= 3000,
+    );
+
+    // We want to do nothing at 1 and not 0 here because this guarantees that
+    // there is at least one timestamp before the sample set. This means that
+    // there are at least three seconds of samples.
+    if (startIndex < 1) {
+      return;
+    }
+
+    const samples = this._volumeStats.values.slice(
+      startIndex > 0
+        ? startIndex
+        : 0,
+    );
+
+    // Calculate the standard deviation of the sample set.
+    const sampleAverage = samples.reduce(
+      (sample: number, partialSum: number) => sample + partialSum,
+      0,
+    ) / samples.length;
+    const diffSquared = samples.map(
+      (sample: number) => Math.pow(sample - sampleAverage, 2),
+    );
+    const stdDev = Math.sqrt(diffSquared.reduce(
+      (sample: number, partialSum: number) => sample + partialSum,
+      0,
+    ) / samples.length);
+
+    // 255 is max volume value; 2.55 is 1% of max
+    const isConstantAudio = stdDev <= 2.55;
+    if (isConstantAudio && sampleAverage <= 2.55) {
+      if (!this.activeWarnings.has(WarningName.LowAudioLevel)) {
+        this.activeWarnings.add(WarningName.LowAudioLevel);
+        this.emit(InputTest.Events.Warning, WarningName.LowAudioLevel);
+      }
+    } else if (this.activeWarnings.has(WarningName.LowAudioLevel)) {
+      this.activeWarnings.delete(WarningName.LowAudioLevel);
+      this.emit(InputTest.Events.WarningCleared, WarningName.LowAudioLevel);
+    }
   }
 
   /**
@@ -414,6 +512,8 @@ export namespace InputTest {
     End = 'end',
     Error = 'error',
     Volume = 'volume',
+    Warning = 'warning',
+    WarningCleared = 'warning-cleared',
   }
 
   /**