+++ lib/constants.ts
@@ -42,7 +42,44 @@ export const BITRATE_TEST_TIMEOUT_MS: number = 15000;
  * @private
  * Test names.
  */
-export enum TestNames {
+export enum TestName {
   InputAudioDevice = 'input-volume',
   OutputAudioDevice = 'output-volume',
+}
+
+/**
+ * All of the expected error names to be thrown by the diagnostics tests.
+ * These names are set in the error objects as the `.name` member.
+ */
+export enum ErrorName {
+  AlreadyStoppedError = 'already-stopped',
+  DiagnosticError = 'diagnostic',
+  InvalidOptionError = 'invalid-option',
+  InvalidOptionsError = 'invalid-options',
+  InvalidStateError = 'invalid-state',
+  PromiseTimedOutError = 'promise-timed-out',
+  UnsupportedError = 'unsupported',
+}
+
+/**
+ * All of the expected warnings to be thrown by the diagnostics tests.
+ */
+export enum WarningName {
+  /**
+   * The `low-audio-level` warning is raised when the volume events recorded
+   * by the input audio device test [[InputTest]] are both low and constant.
+   *
+   * The warning criteria is when the following are all true:
+   * - If there are at least three seconds worth of audio samples.
+   * - The standard deviation of those samples is less than 1% of the max
+   *   possible volume value (255).
+   * - The average of those samples is less than 1% of the max possible volume
+   *   value (255).
+   *
+   * When any of the previous criteria are no longer met, the `warning-cleared`
+   * event for `low-audio-level` is raised if `low-audio-level` has been raised.
+   * Only one `low-audio-level` warning will be raised until the
+   * `warning-cleared` event has been raised.
+   */
+  LowAudioLevel = 'low-audio-level',
 }