+++ lib/errors/DiagnosticError.ts
@@ -1,3 +1,5 @@
+import { ErrorName } from '../constants';
+
 /**
  * @internalapi
  * Generic Diagnostic SDK error that provides a superclass for all other errors.
@@ -7,6 +9,13 @@ export class DiagnosticError extends Error {
    * The associated `DOMError` that caused this `DiagnosticError`.
    */
   domError: DOMError | DOMException | undefined;
+  /**
+   * The name of the error.
+   *
+   * This overrides the typing of the `Error.name` superclass member from
+   * `string` to the specific enumeration of error names.
+   */
+  name: ErrorName;
   /**
    * The timestamp of the occurrence of this error.
    */
@@ -24,6 +33,6 @@ export class DiagnosticError extends Error {
 
     Object.setPrototypeOf(this, DiagnosticError.prototype);
 
-    this.name = 'DiagnosticError';
+    this.name = ErrorName.DiagnosticError;
   }
 }