+++ lib/types.ts
@@ -35,3 +35,27 @@ export interface TimeMeasurement {
  * @internalapi
  */
 export type SubsetRequired<T, K extends keyof T> = T & Required<Pick<T, K>>;
+
+/**
+ * @internalapi
+ * Volume statistics to track in the audio device tests.
+ */
+export interface VolumeStats {
+  /**
+   * The maximum volume value recorded during the test.
+   */
+  max?: number;
+  /**
+   * The smallest non-zero volume value encountered.
+   */
+  min?: number;
+  /**
+   * The timestamps of the recorded values. For example, the 0th indexed
+   * timestamp correlates to when the 0th indexed volume value was recorded.
+   */
+  timestamps: number[];
+  /**
+   * The recorded volume values encountered during the test.
+   */
+  values: number[];
+}